Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

As a base YAML file to get us started this looks fine. Is the automation that processes it ready to go? If so, I would suggest that we merge this as-is and then amend it in order to verify that the automation works.

Patch-set: 2
Reviewer: Gerrit User 43 <43@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Code-Review=+1